python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "modei-sdk"
version = "1.0.0"
description = "Python SDK for the Modei REST API — manage agent passports, gates, and enforcement policies."
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
authors = [
    { name = "Standard Logic Co.", email = "support@modei.ai" },
]
keywords = [
    "modei",
    "ai-agents",
    "permissions",
    "passports",
    "trust",
    "mcp",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Typing :: Typed",
]
dependencies = [
    "httpx>=0.25.0",
    "pydantic>=2.0.0",
    "pynacl>=1.5.0",
    "canonicaljson>=2.0.0",
]

[project.optional-dependencies]
realtime = [
    "supabase>=2.0.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "respx>=0.21.0",
]

[project.urls]
Homepage = "https://modei.ai"
Documentation = "https://modei.ai/guide"
Repository = "https://github.com/StandardLogic/modei-mcp"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.hatch.build.targets.wheel]
packages = ["src/modei"]
